% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_model.R
\name{boot_test}
\alias{boot_test}
\title{Obtain OOB sample to use as test set}
\usage{
boot_test(train.id)
}
\arguments{
\item{train.id}{list of training set indices}
}
\description{
Obtain OOB sample to use as test set
}
