% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conds.R
\name{cond_function_0}
\alias{cond_function_0}
\title{Conditioning 0}
\usage{
cond_function_0(data, l_2)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}
}
\value{
a boolean
}
\description{
Conditioning 0
}
\author{
Giovanni Laudanno
}
