% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_pars.R
\name{create_pars}
\alias{create_pars}
\title{Create parameters}
\usage{
create_pars(lambdas, mus, ks = c(Inf, Inf))
}
\arguments{
\item{lambdas}{speciation rates, for all the clades}

\item{mus}{extinction rate, for all the clades}

\item{ks}{carrying capacities, for all the clades}
}
\value{
parameters
}
\description{
Create parameters
}
\author{
Giovanni Laudanno
}
