% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_convert.R
\name{dummify}
\alias{dummify}
\title{Convert all categorical variables in a dataset to dummy variables}
\usage{
dummify(data)
}
\arguments{
\item{data}{data frame}
}
\value{
A numeric data frame where dummy variables encode categorical
  variables using multiple columns. Continuous variables are unchanged.
}
\description{
Convert all categorical variables in a dataset to dummy variables
}
\author{
Derek Chiu
}
