% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error_632}
\alias{error_632}
\title{.632(+) Estimator for log loss error rate}
\usage{
error_632(data, class, algorithm, pred, test.id, train.id, plus = TRUE)
}
\arguments{
\item{data}{data frame with rows as samples, columns as features}

\item{class}{true/reference class vector used for supervised learning}

\item{algorithm}{character string for classifier. See \code{splendid} for
possible options.}

\item{pred}{vector of OOB predictions using the same classifier as
\code{algorithm}.}

\item{test.id}{vector of test set indices for each bootstrap replicate}

\item{train.id}{vector of training set indices for each bootstrap replicate}

\item{plus}{logical; if \code{TRUE} (default), the .632+ estimtor is
calculated. Otherwise, the .632 estimator is calculated.}
}
\value{
the .632(+) log loss error rate
}
\description{
The .632 estimator for the log loss error rate is calculated for a given
classifer. The .632+ estimator is an extension that reduces overfitting and
is run by default.
}
\details{
This function is intended to be used internally by \code{splendid_model}.
}
\examples{
data(hgsc)
class <- as.factor(attr(hgsc, "class.true"))
set.seed(1)
train.id <- boot_train(data = hgsc, class = class, n = 5)
test.id <- boot_test(train.id = train.id)
mod <- purrr::map(train.id, ~ classification(hgsc[., ], class[.], "xgboost"))
pred <- purrr::pmap(list(mod = mod, test.id = test.id, train.id = train.id),
prediction, data = hgsc, class = class)
error_632(hgsc, class, "xgboost", pred, test.id, train.id, plus = FALSE)
error_632(hgsc, class, "xgboost", pred, test.id, train.id, plus = TRUE)
}
\references{
Friedman, Jerome, Trevor Hastie, and Robert Tibshirani. The
  elements of statistical learning. Vol. 1. New York: Springer series in
  statistics, 2001.

Efron, Bradley and Tibshirani, Robert (1997), "Improvements on
  Cross-Validation: The .632+ Bootstrap Method," Journal of American
  Statistical Association, 92, 438, 548-560.
}
\author{
Derek Chiu
}
