% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Evaluation of prediction performance}
\usage{
evaluation(x, y)
}
\arguments{
\item{x}{actual class labels}

\item{y}{predicted class labels}
}
\value{
A tibble with one row per algorithm, one column per evaluation
  measure plus a column for the algorithm name.
}
\description{
Evaluation of prediction performance on the OOB set is done using various
measure for classification problems.
}
\details{
The currently supported evaluation measures include overall accuracy, average
accuracy across all One-vs-all confusion matrices, and macro-averaged 
precision, recall, and F1-score.
}
\references{
https://github.com/saidbleik/Evaluation/blob/master/eval.R
}
\author{
Derek Chiu
}
