% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_name}
\alias{get_model_name}
\title{Cut word "loglik" from a name}
\usage{
get_model_name(function_name)
}
\arguments{
\item{function_name}{function name}
}
\value{
clean name
}
\description{
Cut word "loglik" from a name
}
\author{
Giovanni Laudanno
}
