% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_info.R
\name{get_param_names}
\alias{get_param_names}
\title{Get the names of the parameters used in the model}
\usage{
get_param_names()
}
\description{
Get the names of the parameters used in the model
}
\author{
Giovanni Laudanno
}
