% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgsc}
\alias{hgsc}
\title{High grade serous cancer data from TCGA used to classify subtypes}
\format{A data frame with 489 rows and 321 columns.}
\usage{
hgsc
}
\description{
There are 489 samples of expression levels measured on 321 genes. Sample IDs
are in the row names, and gene IDs are in the column names.
}
\keyword{datasets}
