% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_2_brts}
\alias{l_2_brts}
\title{Convert l_0 to brts}
\usage{
l_2_brts(l_0, n_0, dropextinct = TRUE)
}
\arguments{
\item{l_0}{the l table, for a single clade}

\item{n_0}{starting number of lineages}

\item{dropextinct}{TRUE if you want to remove the dead species from the
tree. FALSE otherwise.}
}
\value{
branching times
}
\description{
Convert l_0 to brts
}
\author{
Giovanni Laudanno
}
