% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_lik_conds.R
\name{lik_cond_3}
\alias{lik_cond_3}
\title{Lik conditioning 3}
\usage{
lik_cond_3(pars, brts)
}
\arguments{
\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}

\item{brts}{branchin times}
}
\value{
a probability
}
\description{
Lik conditioning 3
}
\author{
Giovanni Laudanno
}
