% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliks.R
\name{loglik_dummy}
\alias{loglik_dummy}
\title{Dummy loglik function}
\usage{
loglik_dummy(brts, pars, n_0, t_0)
}
\arguments{
\item{brts}{branchin times}

\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}

\item{n_0}{starting number of lineages}

\item{t_0}{starting time}
}
\value{
the loglik
}
\description{
Yields a dummy loklik
}
\author{
Giovanni Laudanno
}
