% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ova.R
\name{ova_predict}
\alias{ova_predict}
\title{One-Vs-All prediction approach}
\usage{
ova_predict(fits, data, test.id, class, train.id, threshold = 0.5, ...)
}
\arguments{
\item{fits}{list of ova fits from \code{ova_train}}

\item{data}{data object with rows as samples, columns as features}

\item{class}{true/reference class vector used for supervised learning}

\item{threshold}{a numeric indicating the lowest maximum class probability
below which a sample will be unclassified.}

\item{...}{additional arguments to \code{splendid_model}}
}
\value{
(tibble) predicted probabilities for each class
}
\description{
One-Vs-All prediction approach
}
\author{
Dustin Johnson, Derek Chiu
}
