% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate an sls process}
\usage{
sim(lambdas, mus, ks = c(Inf, Inf), cond = 1,
  l_2 = sim_get_standard_l_2())
}
\arguments{
\item{lambdas}{speciation rates, for all the clades}

\item{mus}{extinction rate, for all the clades}

\item{ks}{carrying capacities, for all the clades}

\item{cond}{type of conditioning:
\itemize{
  \item cond = 0 no conditiong;
  \item cond = 1 conditions on the survival of crown descendents;
  \item cond = 2 not available;
  \item cond = 3 conditions on the survival of subclade and
  on the other crown descendents in the main clade;
  \item cond = 4 conditions on the survival of the subclade and
  both crown descendents in the main clade;
}}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}
}
\value{
l_1 table and brts
}
\description{
Simulate an sls process
}
\author{
Giovanni Laudanno
}
