% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_sim_conds.R
\name{sim_cond_2}
\alias{sim_cond_2}
\title{Sim conditioning 2}
\usage{
sim_cond_2(data, l_2)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}
}
\value{
a boolean
}
\description{
Sim conditioning 2
}
\author{
Giovanni Laudanno
}
