% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils2.R
\name{sim_decide_event2}
\alias{sim_decide_event2}
\title{Determines the event}
\usage{
sim_decide_event2(data, clade, delta_t, l_2)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{clade}{the id of the clade}

\item{delta_t}{in the Doob-Gillespie algorithm,
the waiting time for the next event to occur}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}
}
\value{
the event
}
\description{
sim module
}
\author{
Giovanni Laudanno
}
