% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{sim_event}
\alias{sim_event}
\title{Apply the event to the simulation}
\usage{
sim_event(data, clade, l_2, delta_t)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{clade}{the id of the clade}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}

\item{delta_t}{in the Doob-Gillespie algorithm,
the waiting time for the next event to occur}
}
\value{
the output of the event
}
\description{
sim module
}
\author{
Giovanni Laudanno
}
