% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{sim_get_pars}
\alias{sim_get_pars}
\title{Get parameters}
\usage{
sim_get_pars(lambdas, mus, ks = c(Inf, Inf))
}
\arguments{
\item{lambdas}{speciation rates, for all the clades}

\item{mus}{extinction rate, for all the clades}

\item{ks}{carrying capacities, for all the clades}
}
\value{
parameters
}
\description{
sim module
}
\author{
Giovanni Laudanno
}
