% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{sim_read_l_1}
\alias{sim_read_l_1}
\title{Reads l_1}
\usage{
sim_read_l_1(l_1)
}
\arguments{
\item{l_1}{the collection of all the l tables, for all the clades}
}
\value{
l_1
}
\description{
sim module
}
\author{
Giovanni Laudanno
}
