% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{sim_sample_delta_t}
\alias{sim_sample_delta_t}
\title{Sample deltas for Doob-Gillespie}
\usage{
sim_sample_delta_t(data, clade, pars, l_2)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{clade}{the id of the clade}

\item{pars}{parameters of the likelihood functions:
\itemize{
  \item pars[1] is lambda_m, i.e. speciation rate of the main clade;
  \item pars[2] is mu_m, i.e. extinction rate of the main clade;
  \item pars[3] is lambda_s, i.e. speciation rate of the sub clade;
  \item pars[4] is mu_s, i.e. extinction rate of the sub clade;
}}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}
}
\value{
delta_t
}
\description{
sim module
}
\author{
Giovanni Laudanno
}
