% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils2.R
\name{sim_use_event2}
\alias{sim_use_event2}
\title{Update data and time given the event}
\usage{
sim_use_event2(data, clade, l_2, event, delta_t)
}
\arguments{
\item{data}{contains all the information about the simulated process}

\item{clade}{the id of the clade}

\item{l_2}{the matrix containing the information about how the subclades are
nested into the main clade. See sls_sim.get_standard_l_2() for more info.}

\item{event}{the event occurring in the simulated process at a given time}

\item{delta_t}{in the Doob-Gillespie algorithm,
the waiting time for the next event to occur}
}
\value{
data
}
\description{
sim module
}
\author{
Giovanni Laudanno
}
