% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid-package.r
\docType{package}
\name{splendid-package}
\alias{splendid-package}
\title{splendid}
\description{
The splendid package provides an ensemble framework for classifying,
predicting, and evaluating supervised learning problems. An ensemble is built
across bootstrap resamples as well as across different machine learning
classification algorithms.
}
