% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid-package.r, R/splendid.R
\docType{package}
\name{splendid}
\alias{splendid}
\alias{splendid-package}
\alias{splendid}
\title{splendid.}
\usage{
splendid(data, class, n, seed = 1, algorithms = NULL)
}
\arguments{
\item{data}{data object with rows as samples, columns as features}

\item{class}{reference class used for supervised learning}

\item{n}{number of bootstrap replicates to generate}

\item{seed}{random seed used for reproducibility in bootstrapping results}

\item{algorithms}{character vector of algorithm names to use for supervised 
learning. Currently supports "lda", "rf", "multinom", "nnet", "svm", "pam",
"adaboost", and "nb".}
}
\value{
A nested list with two elements: "model", "pred", and "eval". 
  Elements "model" and "pred" show the models and predictions respectively 
  across the algorithms used and each bootstrap replicate. Element "eval" is 
  a tibble indicating the median of the aggregate evaluation measures for 
  each algorithm and measure used. Evaluation measures include accuracy, 
  average accuracy, average precision, average recall, and average F1-score. 
  The average is taken across the number of classes in \code{class}, and 
  there is one average for ebery bootstrap replicate. We arrive at the final 
  tibble by calculating the median of these measures across replicates, 
  omitting any missing entries.
}
\description{
splendid.

Supervised learning classification algorithms performed on bootstrap 
replicates.
}
\details{
Training sets are bootstrap replicates, and test sets comprise of remaining 
samples not chosen for each training set.
}
\author{
Derek Chiu
}
