% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_convert.R
\name{splendid_convert}
\alias{splendid_convert}
\title{Convert covariate data}
\usage{
splendid_convert(data, algorithms, convert = FALSE)
}
\arguments{
\item{data}{data frame with rows as samples, columns as features}

\item{algorithms}{character vector of algorithms to use for supervised
learning. See \strong{Algorithms} section for possible options. By default,
this argument is \code{NULL}, in which case all algorithms are used.}

\item{convert}{logical; if \code{TRUE}, converts all categorical variables in
\code{data} to dummy variables. Certain algorithms only work with such
limitations (e.g. LDA).}
}
\value{
A (potentially) transformed data frame.
}
\description{
Converts all categorical predictors to dummy variables in the dataset.
Classification algorithms LDA and the MLR family have such a limitation.
}
\details{
If all the variables in the original data are already continuous, nothing is
done. Otherwise, conversion is performed if \code{convert = TRUE}. An error
message is thrown if there are categorical variables and \code{convert =
FALSE}, indicating exactly which algorithms specified require data
conversion.
}
\author{
Derek Chiu
}
