% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_ensemble.R
\name{splendid_ensemble}
\alias{splendid_ensemble}
\title{Combine classification models into an ensemble}
\usage{
splendid_ensemble(sm, data, class, top = 3, seed = 1, rfe = FALSE)
}
\description{
Combine classification models into an ensemble
}
\examples{
data(hgsc)
class <- stringr::str_split_fixed(rownames(hgsc), "_", n = 2)[, 2]
sm <- splendid_model(hgsc, class, n = 2, algorithms = c("svm", "lda"))
se <- splendid_ensemble(sm, hgsc, class)
}
