% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{splendid_graphs}
\alias{splendid_graphs}
\alias{discrimination_plot}
\alias{reliability_plot}
\title{Discriminating graphs}
\usage{
discrimination_plot(x, pred.probs)

reliability_plot(x, pred.probs)
}
\arguments{
\item{x}{true class labels}

\item{pred.probs}{matrix of predicted class probabilities. Number of rows
must equal length of \code{x}}
}
\value{
\code{ggplot} objects for the desired plot
}
\description{
Graphs a \code{discrimination_plot} and \code{reliability_plot} based on true classes
and predicted class probabilities.
}
\details{
A \code{discrimination_plot} shows boxplots of the predicted probabilities for
each class, separated by panels of the true class labels. The class
prevalence is also drawn as a horizontal line for each panel.

A \code{reliability_plot} shows mean prediction vs. observed fraction on lowess
smoother for each class. A line going thru the origin with slope of 1 serves
as a reference for perfect reliability.

Both plots can be called from within \code{\link[=evaluation]{evaluation()}}.
}
\examples{
data(hgsc)
class <- attr(hgsc, "class.true")
set.seed(1)
training.id <- sample(seq_along(class), replace = TRUE)
test.id <- which(!seq_along(class) \%in\% training.id)
mod <- classification(hgsc[training.id, ], class[training.id], "xgboost")
pred <- prediction(mod, hgsc, test.id, class = class)
discrimination_plot(class[test.id], attr(pred, "prob"))
reliability_plot(class[test.id], attr(pred, "prob"))
}
\references{
http://onlinelibrary.wiley.com/doi/10.1002/sim.5321/abstract
}
\author{
Dustin Johnson, Derek Chiu
}
