% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_process.R
\name{subsample}
\alias{subsample}
\title{Subsampling Imbalanced Data}
\usage{
subsample(data, class, sampling = c("none", "up", "down", "smote"),
  seed_samp = NULL)
}
\arguments{
\item{data}{data frame with rows as samples, columns as features}

\item{class}{true/reference class vector used for supervised learning}

\item{sampling}{the default is "none", in which no subsampling is performed.
Other options include "up" (Up-sampling the minority class), "down"
(Down-sampling the majority class), and "smote" (synthetic points for the
minority class and down-sampling the majority class). Subsampling is only
applicable to the training set.}

\item{seed_samp}{random seed used for reproducibility in subsampling
training sets for model generation}
}
\value{
A subsampled dataset where corresponding strata of \code{class} are more
balanced. The resulting \code{class} variable is not included in the data
output.
}
\description{
Subsampling imbalanced data using up-sampling, down-sampling, or SMOTE.
}
\details{
To deal with class imbalances, we can subsample the data so that the class
proportions are more uniform.
}
\examples{
# Create imbalanced version of iris dataset
iris_imbal <- iris[1:130, ]

# Up-sampling
iris_up <- subsample(iris_imbal, iris_imbal$Species, sampling = "up")
nrow(iris_up)

# Down-sampling
iris_down <- subsample(iris_imbal, iris_imbal$Species, sampling = "down")
nrow(iris_down)

# SMOTE
iris_smote <- subsample(iris_imbal, iris_imbal$Species, sampling = "smote")
nrow(iris_smote)
}
\author{
Derek Chiu
}
