% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Machine learning classification algorithms}
\usage{
classification(x, y, algs)
}
\arguments{
\item{x}{data}

\item{y}{class}

\item{algs}{algorithms}
}
\description{
Machine learning classification algorithms
}
\note{
qda errors because there are too many columns; need to reduce so that 
  every class is greater than number of variables
}
\author{
Derek Chiu
}
