% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_process.R
\name{dummify}
\alias{dummify}
\title{Create dummy variables}
\usage{
dummify(data)
}
\arguments{
\item{data}{data frame with rows as samples, columns as features}
}
\value{
A numeric data frame where dummy variables encode categorical
variables using multiple columns. Continuous variables are unchanged.
}
\description{
Convert all categorical variables in a dataset to dummy variables
}
\examples{
dummify(mtcars)
dummify(iris)
}
\author{
Derek Chiu
}
