% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Evaluate class prediction labels}
\usage{
evaluation(x, y)
}
\arguments{
\item{x}{actual class labels}

\item{y}{predicted class labels}
}
\description{
Evaluate class prediction labels
}
\references{
https://github.com/saidbleik/Evaluation/blob/master/eval.R
}
\author{
Derek Chiu
}
