% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ova.R
\name{ova_classification}
\alias{ova_classification}
\title{One-Vs-All training approach}
\usage{
ova_classification(data, class, algorithms, rfe = FALSE, ova = FALSE,
  standardize = FALSE)
}
\arguments{
\item{data}{data frame with rows as samples, columns as features}

\item{class}{true/reference class vector used for supervised learning}

\item{algorithms}{character string of algorithm to use for supervised
learning. See \strong{Algorithms} section for possible options.}

\item{rfe}{logical; if \code{TRUE}, run Recursive Feature Elimination as a
feature selection method for "lda", "rf", and "svm" algorithms.}

\item{ova}{logical; if \code{TRUE}, use the One-Vs-All approach for the
\code{knn} algorithm.}

\item{standardize}{logical; if \code{TRUE}, the training sets are
standardized by per feature to have mean zero and unit variance. The test
sets are standardized using the vectors of centers and standard deviations
used in corresponding training sets.}
}
\value{
list of binary classifier fits on each class
}
\description{
One-Vs-All training approach
}
\author{
Dustin Johnson, Derek Chiu
}
