% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ova.R
\name{ova_prediction}
\alias{ova_prediction}
\title{One-Vs-All prediction approach}
\usage{
ova_prediction(fits, data, test.id = NULL, class, train.id, threshold = 0,
  standardize = FALSE, ...)
}
\arguments{
\item{fits}{list of ova fits from \code{ova_classification}}

\item{data}{data frame with rows as samples, columns as features}

\item{test.id}{integer vector of indices for test set. If \code{NULL} (default),
all samples are used.}

\item{class}{true/reference class vector used for supervised learning}

\item{train.id}{integer vector of indices for training set. If \code{NULL}
(default), all samples are used.}

\item{threshold}{a number between 0 and 1 indicating the lowest maximum class
probability below which a sample will be unclassified.}

\item{standardize}{logical; if \code{TRUE}, the training sets are standardized on
features to have mean zero and unit variance. The test sets are
standardized using the vectors of centers and standard deviations used in
corresponding training sets.}

\item{...}{additional arguments to be passed to or from methods}
}
\value{
(tibble) predicted probabilities for each class
}
\description{
One-Vs-All prediction approach
}
\author{
Dustin Johnson, Derek Chiu
}
