% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predict using mod onto test set indicated by indices in test.id
If model is of class "lda", the predictions are accessed in the "class" name}
\usage{
prediction(mod, train.id, test.id, data, class)
}
\arguments{
\item{mod}{model}

\item{train.id}{training set ID}

\item{test.id}{test set ID}

\item{data}{data}

\item{class}{class}
}
\description{
Predict using mod onto test set indicated by indices in test.id
If model is of class "lda", the predictions are accessed in the "class" name
}
\author{
Derek Chiu
}
