% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Prediction on OOB set}
\usage{
prediction(mod, data, test.id, ...)
}
\arguments{
\item{mod}{model}

\item{data}{data}

\item{test.id}{test set ID}

\item{...}{additional arguments to be passed to or from methods}

\item{train.id}{training set ID}

\item{class}{class}
}
\value{
A factor of predicted values with levels in the same order as true
  class
}
\description{
Functions to predict class labels on the Out-Of-Bag set for different models.
}
\author{
Derek Chiu
}
