% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\alias{prediction.default}
\alias{prediction.pamrtrained}
\alias{prediction.knn}
\title{Class prediction on OOB set}
\usage{
prediction(mod, data, class, test.id = NULL, train.id = NULL,
  threshold = 0, standardize = FALSE, ...)

\method{prediction}{default}(mod, data, class, test.id = NULL,
  train.id = NULL, threshold = 0, standardize = FALSE, ...)

\method{prediction}{pamrtrained}(mod, data, class, test.id = NULL,
  train.id = NULL, threshold = 0, standardize = FALSE, ...)

\method{prediction}{knn}(mod, data, class, test.id = NULL, train.id = NULL,
  threshold = 0, standardize = FALSE, ...)
}
\arguments{
\item{mod}{model object from \code{\link[=classification]{classification()}}}

\item{data}{data frame with rows as samples, columns as features}

\item{class}{true/reference class vector used for supervised learning}

\item{test.id}{integer vector of indices for test set. If \code{NULL} (default),
all samples are used.}

\item{train.id}{integer vector of indices for training set. If \code{NULL}
(default), all samples are used.}

\item{threshold}{a number between 0 and 1 indicating the lowest maximum class
probability below which a sample will be unclassified.}

\item{standardize}{logical; if \code{TRUE}, the training sets are standardized on
features to have mean zero and unit variance. The test sets are
standardized using the vectors of centers and standard deviations used in
corresponding training sets.}

\item{...}{additional arguments to be passed to or from methods}
}
\value{
A factor of predicted classes with labels in the same order as true
class. If \code{mod} is a \code{"pamr"} classifier, the return value is a list of
length 2: the predicted class, and the threshold value.
}
\description{
Functions to predict class labels on the Out-Of-Bag (test) set for different
classifiers.
}
\details{
The \code{knn} and \code{pamr} prediction methods use the \code{train.id} and \code{class}
arguments for additional modelling steps before prediction. For \code{knn}, the
modelling and prediction are performed in one step, so the function takes in
both training and test set identifiers. For \code{pamr}, the classifier needs to
be cross-validated on the training set in order to find a shrinkage threshold
with the minimum CV error to use in prediction on the test set. The other
prediction methods make use of the default method.
}
\examples{
data(hgsc)
class <- attr(hgsc, "class.true")
set.seed(1)
training.id <- sample(seq_along(class), replace = TRUE)
test.id <- which(!seq_along(class) \%in\% training.id)
mod <- classification(hgsc[training.id, ], class[training.id], "slda")
pred <- prediction(mod, hgsc, class, test.id)
table(true = class[test.id], pred)
}
\author{
Derek Chiu
}
