% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid-package.r
\docType{package}
\name{splendid-package}
\alias{splendid-package}
\alias{_PACKAGE}
\title{splendid: SuPervised Learning ENsemble for Diagnostic IDentification}
\description{
Provides a bootstrapping and ensemble framework for supervised learning analyses using multiclass classification algorithms for modelling, prediction, and evaluation. Predicted classes are evaluated under metrics such as log loss, AUC, F1-score, Matthew's correlation coefficient, and accuracy. Discrimination and reliability plots visualize the classifier performances. The .632+ estimator is implemented for the log loss error rate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlineTalhouk/splendid}
  \item \url{https://alinetalhouk.github.io/splendid/}
  \item Report bugs at \url{https://github.com/AlineTalhouk/splendid/issues}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

Authors:
\itemize{
  \item Aline Talhouk \email{atalhouk@bccrc.ca}
  \item Dustin Johnson \email{djohnson@bccrc.ca}
}

}
\keyword{internal}
