% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_ensemble.R
\name{splendid_ensemble}
\alias{splendid_ensemble}
\title{Combine classification models into an ensemble}
\usage{
splendid_ensemble(sm, data, class, top = 3, seed_rank = 1,
  rfe = FALSE, sequential = FALSE)
}
\arguments{
\item{sm}{a \code{splendid_model} object}

\item{data}{data frame with rows as samples, columns as features}

\item{class}{true/reference class vector used for supervised learning}

\item{top}{the number of highest-performing algorithms to retain for ensemble}

\item{seed_rank}{random seed used for reproducibility in rank aggregation of
ensemble algorithms}

\item{rfe}{logical; if \code{TRUE}, run Recursive Feature Elimination as a feature
selection method for "lda", "rf", and "svm" algorithms.}

\item{sequential}{logical; if \code{TRUE}, a sequential model is fit on the
algorithms that had the best performance with one-vs-all classification.}
}
\description{
Combine classification models into an ensemble
}
\examples{
dat <- iris[, 1:4]
class <- iris$Species
sm <- splendid_model(dat, class, n = 3, algorithms = c("xgboost", "slda"))
se <- splendid_ensemble(sm, dat, class)
}
