% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_model.R
\name{splendid_model}
\alias{splendid_model}
\title{Train, predict, and evaluate classification models}
\usage{
splendid_model(data, class, n, seed = 1, algorithms = NULL,
  conf.level = 0.95, rfe = FALSE, ...)
}
\description{
Train, predict, and evaluate classification models
}
\examples{
data(hgsc)
class <- stringr::str_split_fixed(rownames(hgsc), "_", n = 2)[, 2]
sl_result <- splendid_model(hgsc, class, n = 1, algorithms = "svm")
}
