% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_model.R
\name{splendid_model}
\alias{splendid_model}
\title{Train, predict, and evaluate classification models}
\usage{
splendid_model(data, class, n, seed = 1, algorithms = NULL, rfe = FALSE,
  ...)
}
\arguments{
\item{data}{data object with rows as samples, columns as features}

\item{class}{true/reference class vector used for supervised learning}

\item{n}{number of bootstrap replicates to generate}

\item{seed}{random seed used for reproducibility in bootstrapping results}

\item{algorithms}{character vector of algorithm names to use for supervised
learning. See Details for possible options. This argument is \code{NULL} by
default, in which case uses all implemented algorithms.}

\item{rfe}{logical; if \code{TRUE}, run Recursive Feature Elimination as a
feature selection method for "lda", "qda", and "rf" algorithms.}

\item{...}{additional arguments to \code{splendid_model}}
}
\description{
Train, predict, and evaluate classification models
}
\examples{
data(hgsc)
class <- stringr::str_split_fixed(rownames(hgsc), "_", n = 2)[, 2]
sl_result <- splendid_model(hgsc, class, n = 1, algorithms = "svm")
}
