% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splendid_model.R
\name{splendid_model}
\alias{splendid_model}
\title{Train, predict, and evaluate classification models}
\usage{
splendid_model(data, class, algorithms = NULL, n = 1, seed_boot = 1,
  convert = FALSE, rfe = FALSE, ova = FALSE, standardize = FALSE,
  plus = TRUE, threshold = 0, trees = 500, tune = FALSE)
}
\arguments{
\item{data}{data frame with rows as samples, columns as features}

\item{class}{true/reference class vector used for supervised learning}

\item{algorithms}{character vector of algorithms to use for supervised
learning. See \strong{Algorithms} section for possible options. By default,
this argument is \code{NULL}, in which case all algorithms are used.}

\item{n}{number of bootstrap replicates to generate}

\item{seed_boot}{random seed used for reproducibility in bootstrapping
training sets for model generation}

\item{convert}{logical; if \code{TRUE}, converts all categorical variables in
\code{data} to dummy variables. Certain algorithms only work with such
limitations (e.g. LDA).}

\item{rfe}{logical; if \code{TRUE}, run Recursive Feature Elimination as a feature
selection method for "lda", "rf", and "svm" algorithms.}

\item{ova}{logical; if \code{TRUE}, a One-Vs-All classification approach is
performed for every algorithm in \code{algorithms}. The relevant results are
prefixed with the string \code{ova_}.}

\item{standardize}{logical; if \code{TRUE}, the training sets are standardized on
features to have mean zero and unit variance. The test sets are
standardized using the vectors of centers and standard deviations used in
corresponding training sets.}

\item{plus}{logical; if \code{TRUE} (default), the .632+ estimator is calculated.
Otherwise, the .632 estimator is calculated.}

\item{threshold}{a number between 0 and 1 indicating the lowest maximum class
probability below which a sample will be unclassified.}

\item{trees}{number of trees to use in "rf" or boosting iterations (trees) in
"adaboost"}

\item{tune}{logical; if \code{TRUE}, algorithms with hyperparameters are tuned}
}
\description{
Train, predict, and evaluate classification models
}
\section{Algorithms}{
 The classification algorithms currently supported are:
\itemize{
\item Prediction Analysis for Microarrays ("pam")
\item Support Vector Machines ("svm")
\item Random Forests ("rf")
\item Linear Discriminant Analysis ("lda")
\item Shrinkage Linear Discriminant Analysis ("slda")
\item Shrinkage Diagonal Discriminant Analysis ("sdda")
\item Multinomial Logistic Regression using
\itemize{
\item Generalized Linear Model with no penalization ("mlr_glm")
\item GLM with LASSO penalty ("mlr_lasso")
\item GLM with ridge penalty ("mlr_ridge")
\item Neural Networks ("mlr_nnet")
}
\item Neural Networks ("nnet")
\item Naive Bayes ("nbayes")
\item Adaptive Boosting ("adaboost")
\item AdaBoost.M1 ("adaboost_m1")
\item Extreme Gradient Boosting ("xgboost")
\item K-Nearest Neighbours ("knn")
}
}

\examples{
data(hgsc)
class <- attr(hgsc, "class.true")
sl_result <- splendid_model(hgsc, class, n = 1, algorithms = "xgboost")
}
