% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.R
\name{var_imp}
\alias{var_imp}
\alias{var_imp.default}
\alias{var_imp.randomForest}
\alias{var_imp.maboost}
\alias{var_imp.train}
\title{Variable Importance}
\usage{
var_imp(mod, ...)

\method{var_imp}{default}(mod, ...)

\method{var_imp}{randomForest}(mod, ...)

\method{var_imp}{maboost}(mod, ...)

\method{var_imp}{train}(mod, ...)
}
\arguments{
\item{mod}{model object from \code{\link[=classification]{classification()}}}

\item{...}{additional arguments to be passed to or from methods}
}
\description{
Methods to calculate variable importance for different classifiers.
}
\examples{
data(hgsc)
class <- attr(hgsc, "class.true")
mod <- classification(hgsc, class, "xgboost")
var_imp(mod)
}
\author{
Derek Chiu
}
